Refactor to shared code
