+++ tools/screenshots/src/destinations.js
@@ -14,19 +14,7 @@ const DESTINATIONS = {
   },
 };
 
-const execSync = require("../../shared/exec");
-const { sleep } = require("../../shared/util");
-const path = require("path");
-const fs = require("fs-extra");
-const cp = require("child_process");
-const puppeteer = require("puppeteer");
-
-const PROJECT_ROOT = path.resolve(path.join(__dirname, "..", "..", ".."));
-const APP_TO_RUN = "staging-community";
-const VERBOSE = true;
-const HOST = "http://localhost:3000";
-const HEADLESS = false;
-const PAGESIZE = { width: 1200, height: 800 };
+const { Service, Browser } = require("./puppet");
 
 module.exports.cmd = async function (vargs) {
   console.log("Destinations:");
@@ -34,7 +22,7 @@ module.exports.cmd = async function (vargs) {
   for (const key of Object.keys(DESTINATIONS)) {
     try {
       console.log(key);
-      const runner = new Service(key);
+      const runner = new Runner(key);
       await runner.run();
       results.push({ key });
     } catch (error) {
@@ -52,190 +40,39 @@ module.exports.cmd = async function (vargs) {
   }
 };
 
-class Service {
+class Runner {
   constructor(serviceKey) {
     this.data = DESTINATIONS[serviceKey];
     if (!this.data) {
       throw new Error(`Unknown destination: ${serviceKey}`);
     }
-    this.rooCmd = path.join(PROJECT_ROOT, "cli", "dist", "grouparoo.js");
-    this.cwd = path.join(PROJECT_ROOT, "apps", APP_TO_RUN);
-    this.dirPath = path.join(
-      PROJECT_ROOT,
-      "tools",
-      "screenshots",
-      "output",
-      "destinations",
-      serviceKey
-    );
-    this.server = null;
-    this.page = null;
-    this.browser = null;
+    this.service = new Service(this.data.demo, ["destinations", serviceKey]);
   }
 
   async run() {
-    try {
-      await this.open();
-      await this.demo();
-      await this.config();
-      for (const destKey of Object.keys(this.data.destinations)) {
-        const dest = new Destination(this, destKey);
-        await dest.screenshots();
-      }
-    } finally {
-      await this.close();
-    }
-  }
-
-  async open() {
-    if (fs.existsSync(this.dirPath)) {
-      fs.rmSync(this.dirPath, { recursive: true, force: true });
-    }
-  }
-
-  async demo() {
-    const args = this.data.demo;
-    if (!args) return;
-
-    const demo = await execSync(`${this.rooCmd} demo -c ${args}`, {
-      cwd: this.cwd,
-      log: true,
-    });
-    if (VERBOSE) console.log(demo);
-
-    const validate = await execSync(`${this.rooCmd} validate`, {
-      cwd: this.cwd,
-      log: true,
-    });
-    if (VERBOSE) console.log(validate);
-  }
-
-  async config() {
-    return new Promise((done, failed) => {
-      const start = cp.spawn(this.rooCmd, ["config"], { cwd: this.cwd });
-      this.server = start;
-      start.stdout.on("data", function (data) {
-        const output = data.toString();
-        if (output.indexOf("Opening Grouparoo Config in web browser") >= 0) {
-          console.log(output);
-          done();
-        } else if (VERBOSE) console.log(output);
-      });
-
-      start.stderr.on("data", function (data) {
-        console.error("ERROR!");
-        console.error(data.toString());
-        failed();
-      });
-
-      start.on("exit", function (code) {
-        if (VERBOSE) {
-          console.log("process exited with code " + code);
-        }
-      });
-    });
-  }
-
-  async close() {
-    if (this.browser) {
-      await this.browser.close();
-    }
-    this.page = null;
-    if (this.server) {
-      this.server.kill();
-      let count = 0;
-      while (this.server.exitCode === null && count < 1000) {
-        await sleep(10);
-        count++;
-      }
-      this.server = null;
-    }
-  }
-
-  async goto(path) {
-    if (!this.browser) {
-      this.browser = await puppeteer.launch({
-        headless: HEADLESS,
-        defaultViewport: PAGESIZE,
-      });
-    }
-    if (!this.page) {
-      this.page = await this.browser.newPage();
-    }
-    const url = `${HOST}${path}`;
-    await this.page.goto(url);
-    // no alerts
-    await this.page.waitForFunction(
-      () => !document.querySelector(".alert-warning")
-    );
-  }
-}
-
-class Destination {
-  constructor(service, destKey) {
-    this.service = service;
-    this.data = service.data.destinations[destKey];
-    this.shotDir = path.join(service.dirPath, destKey);
-  }
-
-  async goto(path) {
-    return this.service.goto(path);
-  }
-
-  async form(fields) {
-    const page = this.service.page;
-    for (const field of Object.keys(fields)) {
-      const value = fields[field];
-      const selector = `input[name="${field}"]`;
-      const input = await page.waitForSelector(selector);
-      // await page.$eval(selector, (el) => (el.value = value));
-      await input.type(value);
-    }
-
-    await page.click('button[type="submit"]');
-    await page.waitForNavigation();
-  }
-
-  async pickModel() {
-    const page = this.service.page;
-    // pick full model
-    await page.select("#form select", "__model");
-    // wait for preview
-    await page.waitForXPath("//h5[contains(text(), 'Sample Record')]");
-  }
-
-  async screenshot(name, options = {}) {
-    const page = this.service.page;
-    const filePath = path.join(this.shotDir, `${name}.png`);
-    const dirPath = path.dirname(filePath);
-    const selector = options.selector;
-    fs.mkdirpSync(dirPath);
-    if (!selector) {
-      await page.screenshot({ path: filePath, fullPage: true });
-    } else {
-      const border = options.border || 0;
-      await page.waitForSelector(selector);
-      const element = await page.$(selector);
-      const box = await element.boundingBox(); // { x, y, width, height }
-      const x = box.x - border;
-      const y = box.y - border;
-      const width = box.width + border * 2;
-      const height = box.height + border * 2;
-      await page.screenshot({
-        path: filePath,
-        clip: { x, y, width, height },
-      });
-    }
+    await this.service.run(this.screenshots.bind(this));
   }
 
   async screenshots() {
-    const appPath = `/app/${this.service.data.app}`;
-    const destPath = `/model/${this.data.model}/destination/${this.data.id}`;
-    const recordPath = `/model/${this.data.model}/record/new`;
+    const app = new App(this.service, this.data);
+    await app.screenshots();
 
-    // make sure there is a preview
-    await this.goto(recordPath);
-    await this.form({ value: "12" });
+    for (const destKey of Object.keys(this.data.destinations)) {
+      const data = this.data.destinations[destKey];
+      const dest = new Destination(this.service, destKey, data);
+      await dest.screenshots();
+    }
+  }
+}
+
+class App extends Browser {
+  constructor(service, appData) {
+    super(service, ["app"]);
+    this.data = appData;
+  }
+
+  async screenshots() {
+    const appPath = `/app/${this.data.app}`;
 
     // app options
     await this.goto(`${appPath}/edit`);
@@ -244,6 +81,30 @@ class Destination {
       selector: "#appOptions",
       border: 10,
     });
+  }
+}
+
+class Destination extends Browser {
+  constructor(service, destKey, destData) {
+    super(service, [destKey]);
+    this.data = destData;
+  }
+
+  async pickModel() {
+    const page = await this.getPage();
+    // pick full model
+    await page.select("#form select", "__model");
+    // wait for preview
+    await page.waitForXPath("//h5[contains(text(), 'Sample Record')]");
+  }
+
+  async screenshots() {
+    const destPath = `/model/${this.data.model}/destination/${this.data.id}`;
+    const recordPath = `/model/${this.data.model}/record/new`;
+
+    // make sure there is a preview
+    await this.goto(recordPath);
+    await this.form({ value: "12" });
 
     // destination options
     await this.goto(`${destPath}/edit`);
